To return the number of elements in a linked list

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -8,7 +8,7 @@
 
 size_t list_len(const list_t *h)
 {
-	size_t count = 0;
+	unsigned int count = 0;
 
 	while (h != NULL)
 	{
